Feature/#7 (#8)
* refactor: separate each list

* feat: adding animation

+++ lib/models/task_model.dart
@@ -1,7 +1,7 @@
 import './task_type.dart';
 
 class Task {
-  final int id;
+  final BigInt id;
   final String title;
   TaskType type;
   bool done;
